Move parentInfo in seperate file.
