+++ src/modules/files-storage/domain/interface/index.ts
@@ -2,5 +2,6 @@ export * from './file-record-status.interface';
 export * from './file-record.repo.interface';
 export * from './file-storage-parent-type.enum';
 export * from './interfaces';
+export * from './parent-info.interface';
 export * from './preview-width.enum';
 export * from './storage-location.enum';
