+++ src/modules/files-storage/domain/interface/parent-info.interface.ts
@@ -1,0 +1,8 @@
+import { EntityId } from '@shared/domain/types';
+import { ParentReference } from '../file-record.do';
+import { StorageLocation } from './storage-location.enum';
+
+export interface ParentInfo extends ParentReference {
+	storageLocationId: EntityId;
+	storageLocation: StorageLocation;
+}
